+++ src/Controllers/GlobalData/TaxRate.php
@@ -8,14 +8,11 @@ namespace JtlWooCommerceConnector\Controllers\GlobalData;
 
 use jtl\Connector\Model\Identity;
 use jtl\Connector\Model\TaxRate as TaxRateModel;
-use JtlWooCommerceConnector\Controllers\Traits\PullTrait;
 use JtlWooCommerceConnector\Utilities\Db;
 use JtlWooCommerceConnector\Utilities\SqlHelper;
 
 class TaxRate
 {
-    use PullTrait;
-
     public function pullData()
     {
         $return = [];