+++ src/Controllers/Order/CustomerOrderItem.php
@@ -279,16 +279,20 @@ class CustomerOrderItem extends BaseController
 
                 if ($total != 0) {
 
-                    $priceGross = $total + $totalTax;
+                    $detailedTax = $totalTax;
+                    if (isset($taxes['total']) && is_array($taxes['total'])) {
+                        $detailedTax = array_sum($taxes['total']);
+                    }
 
-                    $vat = $this->calculateVat($total, $priceGross, wc_get_price_decimals());
+                    $detailedPriceGross = $total + $detailedTax;
+                    $vat = $this->calculateVat($total, $detailedPriceGross, wc_get_price_decimals());
                     if (!is_null($singleVatRate)) {
                         $vat = $singleVatRate;
                     }
 
                     $customerOrderItem->setVat($vat)
                         ->setPrice(round($total, Util::getPriceDecimals()))
-                        ->setPriceGross(round($priceGross, Util::getPriceDecimals()));
+                        ->setPriceGross(round($detailedPriceGross, Util::getPriceDecimals()));
                 }
 
                 if ($type === CustomerOrderItemModel::TYPE_SHIPPING && $customerOrderItem->getVat() === 0. && $highestVatRateFallback !== 0.) {