+++ src/Controllers/ProductStockLevel.php
@@ -7,13 +7,10 @@
 namespace JtlWooCommerceConnector\Controllers;
 
 use jtl\Connector\Model\ProductStockLevel as ProductStockLevelModel;
-use JtlWooCommerceConnector\Controllers\Traits\PushTrait;
 use JtlWooCommerceConnector\Utilities\Util;
 
 class ProductStockLevel extends BaseController
 {
-    use PushTrait;
-
     public function pushData(ProductStockLevelModel $productStockLevel)
     {
         $productId = $productStockLevel->getProductId()->getEndpoint();